+++ FireStarter/FireStarterProgram.h
@@ -43,8 +43,8 @@ public:
 
     bool Packetize(FireStarterPacket& packet);
     void OptimizeRegisters(bool clean);
-    void RandomProgram(unsigned int seed);
-    void RandomInstruction(unsigned int seed);
+    void RandomProgram(unsigned long long seed);
+    void RandomInstruction(unsigned long long seed);
     void LoadInstructions(FireStarterInstructions* instructions);
     void SaveInstructions(FireStarterInstructions* instructions);
     static void SettingsText(const FireStarterSettings& settings, std::string& code, const std::string& prefix = "");
